feat: fixes reversion of NodePort in ServiceSpec
At present NodePort is not configurable in objects of the
v1alpha1.ServiceSpec type - a reversion from the functionality in v0.3.0.
This leads to issues for sites wishing to use the NodePort attribute of the
corev1.ServicePort API spec. This commit makes the NodePort attribute of
the v1alpha1.ServiceSpec object configurable while preserving the
default behavior when NodePort is undefined (random selection as
needed).

Changelog: added - Added NodePort to v1alpha1.ServiceSpec