docs: fixes cert-manager installation instruction
Installation of cert-manager using the command provided in
installation.md currently fails due to --version not being specified.
The cert-manager docs reccomend the specification of the --version flag
when installing cert-manager, and the specification of this flag
resolves the issues with installation. This commit updates the command
provided for the installation of cert-manager to the functional version.

Closes #215

Changelog: fixed - Fixed installation instruction for cert-manager chart