fix: slurm-controller slurmctldhost for gke dns
Running 0.4.0 in GKE exposed an issue whereby controller.PrimaryFQDN()
was unresolvable, causing the slurmctld to fail. This was due to
controller.PrimaryFQDN() not matching the FQDN of the slurm-controller
service: controller.ServiceFQDNShort(). It is unknown at this time why
this issue was only apparent when running in GKE. This commit replaces
all instances of controller.PrimaryFQDN() with
controller.ServiceFQDNShort() and removes the controller.PrimaryFQDN()
method.

Changelog: fixed - Fixes bug wereby slurm-controller hostname was set
 incorrectly