feat(reconfigure): make more responsive
It is still at the mercy of the Kubernetes API server and Kubelet to
propagate the updated file into the pod. There is still a noticable
delay from when a ConfigMap change is applied and when the mounted
ConfigMap changes in the pod. But when it does, the reconfigure
sidecar will detect it more quickly and issue a reconfigure faster.

Changelog: Changed - Reduced interval of the reconfigure check. After
 the kubelet updates mounted files in the pod, a reconfigure will be
 issued more quickly.