fix: configure loginset slurmEtcVolume for SlurmUser access
Previously sackd within LoginSet pods was executed as SlurmdUser. This is
problematic because presently, sackd must be executed as SlurmUser
(default "slurm"). In Slurm 25.11.0, this is enforced at line 221 of
src/slurmctld/sackd_mgr.c. This fix requires that the permissions in the
slurmEtcVolume permit the user 'slurm' to rwx /etc/slurm. To accomplish
this, the initConf container model has been reintroduced

Changelog: fixed - Configure LoginSet /etc/slurm for SlurmUser access