fix: avoid storage and claim null values
There seems to be a bug with Kubernetes, and/or Helm, where the chart
default/implicit values.yaml will fail when `storageClassName=null`,
despite being an acceptable value. When explicitly providing a
values.yaml with `storageClassName=null`, this works as expected.

$ helm -n slurm install slurm helm/slurm                      # fails
$ helm -n slurm install slurm helm/slurm -f helm/values.yaml  # succeeds

Making existingClaim="" avoid the same null error as storageClassName.
But for storageClassName in particular, null != "", hence we must
specially handle the truthy and falsy with respect to the empty string
and omit when null.

Changelog: Fixed - Avoid `storageClassName=null` error, which occurs when
 using the Slurm chart default `values.yaml`, implicitly used by helm.
Cherry-picked: ac05f32a5c