feat(operator): set NodeSet pod env POD_TOPOLOGY
This is not perfect. There is a race condition between the operator
updating the pod's annotations and the pod's slurmd container starting.
If fast enough, the slurmd will register with its topology per env,
otherwise the operator will update it in the next reconcile.

Changelog: Added - The NodeSet pod's env will now contain POD_TOPOLOGY,
 which allows the slurmd container to start with a topology.